search-api/requirements.txt
perplexityai
python-dotenv
